Changes for VIX-348. Created an auto save menu item under Sequence that allows auto save to be turned on or off. Interval is set to 5 minutes for now. Need to create a general configuration dialog where items such as the save interval can be configured. Will create another enhancement bug for that work as it is bigger than this item. Auto save on/off setting will be saved and restored over editor sessions. Interval is also saved in preparation for configuration dialog. Attributes are saved in the Settings.xml in the user profile that is already purposed.

+++ Common/Controls/XMLProfileSettings.cs
@@ -38,7 +38,7 @@ namespace Common.Controls
 
 		public int GetSetting(string xPath, int defaultValue)
 		{
-			return Convert.ToInt16(GetSetting(xPath, Convert.ToString(defaultValue)));
+			return Convert.ToInt32(GetSetting(xPath, Convert.ToString(defaultValue)));
 		}
 
 		public void PutSetting(string xPath, int value)
